Move to generators directory
