+++ src/main/java/lt/biip/auth/AuthRequestController.java
@@ -1,6 +1,6 @@
 package lt.biip.auth;
 
-import lt.epaslaugos.test.auth.AuthRequestGenerator;
+import lt.biip.auth.generators.AuthRequestGenerator;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
