+++ src/main/java/lt/biip/auth/generators/AuthDataRequestGenerator.java
@@ -1,4 +1,4 @@
-package lt.epaslaugos.test.auth;
+package lt.biip.auth.generators;
 
 import lt.epaslaugos.authentication.client.AuthenticationDataRequest;
 import org.w3c.dom.Document;
